Address CodeRabbit review findings on save FS and ACC stubs.
Fix auto-create attribute checks, cache save indexing, VFS rename cache invalidation, DataErase ordering, FSP filter construction, guest session IDs, async completion races, and web callback URL bounds.

+++ src/core/hle/service/am/service/application_functions.cpp
@@ -341,7 +341,7 @@ Result IApplicationFunctions::CreateCacheStorage(Out<u32> out_target_media,
     FileSys::SaveDataAttribute attribute{};
     attribute.program_id = m_applet->program_id;
     attribute.type = FileSys::SaveDataType::Cache;
-    attribute.index = static_cast<u8>(index);
+    attribute.index = index;
 
     FileSys::VirtualDir save_data{};
     R_TRY(GetApplicationSaveDataController(system, *m_applet)->CreateSaveData(
